app/requirements.txt
asgiref==3.6.0
astropy==5.2
attrs==22.2.0
black==22.12.0
boto3==1.26.35
botocore==1.29.35
click==8.1.3
debugpy==1.5.1
Django==4.1.4
djangorestframework==3.14.0
exceptiongroup==1.1.0
flake8==6.0.0
iniconfig==2.0.0
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.24.0
packaging==22.0
pathspec==0.10.3
platformdirs==2.6.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyerfa==2.0.0.1
pyflakes==3.0.1
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
